Optimization
This is quick optimization of generator script. Current implementation checks every pixel of the image if it is inside of bounds of the polygons (triangles of Ethereum logo) and for each triangle checks each pixel. So for image of 2245 x 1620 it will make ~ 14.5M ckecks (4 x 2245 x 1620).
This change checks only pixels that are inside of the bounds of the triangles, so for same image it will make less then 450K checks (only 3%) and takes 10-15 sec. to create image, compared to 5-7min before.